feature/fsel-5872: update DailyQuizCommand

+++ Services/Fsel.System/Fsel.System.Application/Commands/DailyQuiz/DailyQuizCommand.cs
@@ -243,7 +243,10 @@ namespace Fsel.System.Application.Commands.DailyQuiz
                 {
                     var deleteWinners = await _dailyQuizWinnerRepository.Queryable.Where(p => p.CreatedUserId == _authContext.CurrentUserId).ToListAsync(cancellationToken);
                     deleteWinners = deleteWinners.Where(p => p.CreatedDate.ConvertTimeFromUtc(EnumCountryKey.Vietnam).Date == currentDate.Date).ToList();
-                    await _dailyQuizWinnerRepository.DeleteListAsync(deleteWinners);
+                    if (deleteWinners.Any())
+                    {
+                        await _dailyQuizWinnerRepository.DeleteListAsync(deleteWinners);
+                    }
                 }
 
                 await _dailyQuizHistoryRepository.BulkMergeAsync(dailyQuizHistories, x =>
